% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cases_to_infections.R
\name{cases_to_infections}
\alias{cases_to_infections}
\title{Cases to infections}
\usage{
cases_to_infections(case_data, infection_to_onset, onset_to_report)
}
\arguments{
\item{case_data}{Time series of case incidence data}

\item{infection_to_onset}{PDF infection-to-onset}

\item{onset_to_report}{PDF onset-to-report}
}
\description{
This function allows you to express your love of cats.
}
\examples{
cases_to_infections()
}
\keyword{cases}
