% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccfr_uncertainty.R
\name{ccfr_uncertainty}
\alias{ccfr_uncertainty}
\title{Calculates the maximum likelihood estimate and 95\% confidence interval
of a corrected CFR, using the total cases and total cases with known
outcomes, where the latter replaces the total number of deaths in the
standard (naive) CFR definition. We use a binomial likelihood, approximated
by a Poisson likelihood for large samples}
\usage{
ccfr_uncertainty(total_cases, total_deaths, u_t)
}
\arguments{
\item{total_cases}{The total number of cases observed over the period of an
outbreak of interest. The total number of cases must be greater than or equal
to the total number of deaths}

\item{total_deaths}{The total number of deaths observed over the period of an
outbreak of interest. The total number of deaths must be less than or equal to
the total number of cases}

\item{u_t}{The proportion of cases to cases with known outcomes up to the
point of an outbreak of interest. Used to correct the total number of deaths
for delays between case detection and outcome. Given that its a proportion,
it must be between 0 and 1}
}
\value{
A data.frame containing the MLE estimate and 95\% confidence interval
of the corrected CFR
}
\description{
Calculates the maximum likelihood estimate and 95\% confidence interval
of a corrected CFR, using the total cases and total cases with known
outcomes, where the latter replaces the total number of deaths in the
standard (naive) CFR definition. We use a binomial likelihood, approximated
by a Poisson likelihood for large samples
}
\examples{
ccfr_uncertainty(1000, 900, 0.2)
}
