% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfr-package.R
\docType{package}
\name{cfr-package}
\alias{cfr}
\alias{cfr-package}
\title{cfr: Estimate Disease Severity and Case Ascertainment}
\description{
Estimate the severity of a disease and ascertainment of cases, as discussed in Nishiura et al. (2009) \doi{10.1371/journal.pone.0006852}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/epiverse-trace/cfr}
  \item \url{https://epiverse-trace.github.io/cfr/}
  \item Report bugs at \url{https://github.com/epiverse-trace/cfr/issues}
}

}
\author{
\strong{Maintainer}: Pratik R. Gupte \email{pratik.gupte@lshtm.ac.uk} (\href{https://orcid.org/0000-0001-5294-7819}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Adam Kucharski \email{adam.kucharski@lshtm.ac.uk} (\href{https://orcid.org/0000-0001-8814-9421}{ORCID}) [copyright holder]
  \item Tim Russell \email{timothy.russell@lshtm.ac.uk} (\href{https://orcid.org/0000-0001-5610-6080}{ORCID}) [copyright holder]
}

Other contributors:
\itemize{
  \item Joshua W. Lambert \email{joshua.lambert@lshtm.ac.uk} (\href{https://orcid.org/0000-0001-5218-3046}{ORCID}) [reviewer]
  \item Hugo Gruson \email{hugo.gruson@data.org} (\href{https://orcid.org/0000-0002-4094-1476}{ORCID}) [reviewer]
  \item Tim Taylor \email{tim.taylor@hiddenelephants.co.uk} (\href{https://orcid.org/0000-0002-8587-7113}{ORCID}) [reviewer]
  \item James M. Azam \email{james.azam@lshtm.ac.uk} (\href{https://orcid.org/0000-0001-5782-7330}{ORCID}) [reviewer]
  \item Abdoelnaser M. Degoot \email{abdoelnaser-mahmood.degoot@lshtm.ac.uk} (\href{https://orcid.org/0000-0001-8788-2496}{ORCID}) [reviewer]
  \item Sebastian Funk \email{sebastian.funk@lshtm.ac.uk} (\href{https://orcid.org/0000-0002-2842-3406}{ORCID}) [reviewer]
}

}
\keyword{internal}
