% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfr_calculation.R
\name{cfr_calculation}
\alias{cfr_calculation}
\title{Case fatality risk calculation}
\usage{
cfr_calculation(case_times, case_data, death_data, onset_to_death)
}
\arguments{
\item{case_times}{Time values corresponding to case incidence}

\item{case_data}{Time series of case incidence data}

\item{death_data}{Time series of death incidence data, with same time scale as cases}

\item{onset_to_death}{Probability mass function for onset-to-death}
}
\description{
Function to estimate CFR from case and death data.
}
\examples{
man/examples/cfr_calculation_example
cfr_calculation()
}
\keyword{cases}
