% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_static.R
\name{cfr_static}
\alias{cfr_static}
\title{Estimate a static disease severity measure}
\usage{
cfr_static(data, epidist = NULL, poisson_threshold = 100)
}
\arguments{
\item{data}{A \verb{<data.frame>} containing the outbreak data. A daily time
series with dates or some other absolute indicator of time (e.g. epiday or
epiweek) and the numbers of new cases and new deaths at each time point.
Note that the required columns are "date" (for the date), "cases" (for the
number of reported cases), and "deaths" (for the number of reported deaths)
on each day of the outbreak.

Note that the \verb{<data.frame>} is required to have an unbroken sequence of
dates with no missing dates in between. The "date" column must be of class
\code{Date} (see \code{\link[=as.Date]{as.Date()}}).

Note also that the total number of cases must be greater than the total
number of reported deaths.}

\item{epidist}{An optional argument for the delay distribution used, in the
form of an \verb{<epidist>} object.
Defaults to \code{NULL}, in which case no delay correction is applied.
Passing an \verb{<epidist>} automatically applies delay correction in the
severity estimation.}

\item{poisson_threshold}{The case count above which to use Poisson
approximation. Set to 100 by default.}
}
\value{
A \verb{<data.frame>} with the maximum likelihood estimate and 95\%
confidence interval of the severity estimates, named "severity_mean",
"severity_low", and "severity_high".
}
\description{
Calculates the severity of a disease, while optionally
correcting for reporting delays using an epidemiological delay distribution
of the time between symptom onset and death (onset-to-death).

Other delay distributions may be passed to calculate different disease
severity measures such as the hospitalisation fatality risk.
}
\section{Details: Adjusting for delays between two time series}{
The method used in \code{cfr_static()} follows Nishiura et al.
(2009).
The function calculates a quantity \eqn{u_t} for each day within the input
data, which represents the proportion of cases with a known outcome on day
\eqn{t}.
Following Nishiura et al., \eqn{u_t} is calculated as:
\deqn{u_t = \dfrac{\sum_{i = 0}^t
        \sum_{j = 0}^\infty c_i f_{j - i}}{\sum_{i = 0} c_i}}
where \eqn{f_t} is the value of the probability mass function at time \eqn{t}
and \eqn{c_t}, \eqn{d_t} are the number of new cases and new deaths at time
\eqn{t}, (respectively).
We then use \eqn{u_t} at the end of the outbreak in the following likelihood
function to estimate the severity of the disease in question.
\deqn{{\sf L}({\theta \mid y}) = \log{\dbinom{u_tC}{D}} + D \log{\theta} +
  (u_tC - D)\log{(1.0 - \theta)}}
\eqn{C} and \eqn{D} are the cumulative number of cases and deaths
(respectively) up until time \eqn{t}.
\eqn{\theta} is the parameter we wish to estimate, the severity of the
disease. We estimate \eqn{\theta} using simple maximum-likelihood methods,
allowing the functions within this package to be quick and easy tools to use.

The precise severity measure — CFR, IFR, HFR, etc — that \eqn{\theta}
represents depends upon the input data given by the user.

The epidemiological delay distribution passed to \code{epidist} is used to obtain
a probability mass function parameterised by time; i.e. \eqn{f(t)} which
gives the probability a case has a known outcomes (usually, death) at time
\eqn{t}, parameterised with disease-specific parameters before it is supplied
here.
}

\examples{
# load package data
data("ebola1976")

# get an onset to death distribution from the {epiparameter} package
onset_to_death_ebola <- epiparameter::epidist_db(
  disease = "Ebola Virus Disease",
  epi_dist = "onset_to_death",
  author = "The-Ebola-Outbreak-Epidemiology-Team",
  single_epidist = TRUE
)

# estimate severity without correcting for delays
cfr_static(ebola1976)

# estimate severity while correcting for delays
cfr_static(
  ebola1976,
  epidist = onset_to_death_ebola
)

}
\references{
Nishiura, H., Klinkenberg, D., Roberts, M., & Heesterbeek, J. A. P. (2009).
Early Epidemiological Assessment of the Virulence of Emerging Infectious
Diseases: A Case Study of an Influenza Pandemic. PLOS ONE, 4(8), e6852.
\doi{10.1371/journal.pone.0006852}
}
