% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadelay-package.R
\docType{package}
\name{datadelay-package}
\alias{datadelay}
\alias{datadelay-package}
\title{datadelay: Delay Functions for Epidemic Data}
\description{
Delay functions for routine surveillance data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/epiverse-trace/datadelay}
  \item \url{https://epiverse-trace.github.io/datadelay/}
  \item Report bugs at \url{https://github.com/epiverse-trace/datadelay/issues}
}

}
\author{
\strong{Maintainer}: Adam Kucharski \email{adam.kucharski@lshtm.ac.uk} (\href{https://orcid.org/0000-0001-8814-9421}{ORCID})

Authors:
\itemize{
  \item Tim Russell \email{timothy.russell@lshtm.ac.uk} (\href{https://orcid.org/0000-0001-5610-6080}{ORCID})
  \item Pratik Gupte \email{pratik.gupte@lshtm.ac.uk} (\href{https://orcid.org/0000-0001-5294-7819}{ORCID})
}

}
\keyword{internal}
