% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data.R
\name{.is_pkg_installed}
\alias{.is_pkg_installed}
\title{Check whether a package is installed}
\usage{
.is_pkg_installed(x)
}
\arguments{
\item{x}{A package name as a string.}
}
\value{
Quietly returns a logical for whether the package passed in \code{x} is
available locally.
}
\description{
A small internal function to check whether a required package
is installed.
This function is primarily intended to be used in mocking when testing for
appropriate error messages from functions that rely on a package without
that package being imported in {cfr}.
The only current use case is \code{\link[=prepare_data.incidence2]{prepare_data.incidence2()}}.
}
\keyword{internal}
