% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebola1976.R
\docType{data}
\name{ebola1976}
\alias{ebola1976}
\title{Ebola 1976 outbreak case data}
\format{
\subsection{\code{ebola1976}}{

A data frame with 73 rows and 60 columns:
\describe{
\item{date}{Calendar date}
\item{cases}{Number of cases reported}
\item{deaths}{Number of deaths reported}
}
}
}
\source{
WIP UNKNOWN
}
\usage{
ebola1976
}
\description{
An example epidemic outbreak dataset for use with the \code{datadelay} package.
}
\keyword{datasets}
