% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_ascertainment.R
\name{estimate_ascertainment}
\alias{estimate_ascertainment}
\title{Estimate the ascertainment rate of a disease given a baseline (assumed to be
true) severity estimate}
\usage{
estimate_ascertainment(
  data,
  epidist = NULL,
  type = c("static", "varying"),
  severity_baseline = 0.014,
  burn_in_value = get_default_burn_in(epidist),
  smooth_inputs = FALSE,
  smoothing_window = 1,
  correct_for_delays = FALSE,
  max_date = NULL
)
}
\arguments{
\item{data}{A data.frame containing the outbreak data. A daily time series
with dates or some other absolute indicator of time (e.g. epiday/epiweek) and
the numbers of new cases and new deaths at each time point.
Note that the required columns are "date" (for the date), "cases" (for the
number of reported cases), and "deaths" (for the number of reported deaths)
on each day of the outbreak.
Note that the data.frame is required to have an unbroken sequence of dates
with no missing dates in between. The "date" column must be of the class
\code{Date} (see \code{\link[=as.Date]{as.Date()}}).
Note also that the total number of cases must be greater than the total
number of reported deaths.}

\item{epidist}{The delay distribution used, in the form of an
\code{\link[epiparameter:epidist]{epiparameter::epidist()}} object. This is used to obtain a probability
mass function parameterised by time; i.e. \eqn{f(t)} which gives the
probability a case has a known outcomes (i.e. death) at time \eqn{t},
parameterised with disease-specific parameters before it is supplied here.
A typical example would be a symptom onset to death delay distribution.}

\item{type}{A string, either \code{"static"} or \code{"varying"} which determines
whether \code{\link[=estimate_static]{estimate_static()}} or \code{\link[=estimate_time_varying]{estimate_time_varying()}} is used to calculate
the resulting ascertainment rate}

\item{severity_baseline}{The assumed to be true baseline severity estimate
used in the final ratio to estimate the overall ascertainment rate}

\item{burn_in_value}{The number of time-points (typically days) to disregard
at the start of the time-series, if a burn-in period is desired.
The default value is set to the mean of the central spread of the \verb{<epidist>}
object passed to the function, assuming the temporal resolution is daily.
Alternatively, a sensible value might be 7, to disregard the first week of
cases and deaths.
To consider all case data including the start of the time-series, set this
argument to 1.}

\item{smooth_inputs}{A boolean flag determining whether the user wishes to
smooth the case and death time-series, using a moving average procedure
before calculating the time-varying severity. Useful for noisy time-series
or time-series with strong reporting (e.g., weekend) effects}

\item{smoothing_window}{An \emph{odd} number determining the smoothing window size
to use when smoothing the case and death time-series, using a moving average
procedure before calculating the time-varying severity.
Useful for noisy time-series or time-series with strong reporting
(e.g., weekend) effects.
The default value is 1 for \emph{no smoothing}. Values > 1 apply smoothing.}

\item{correct_for_delays}{A boolean flag indicating whether the user wishes
to correct for the delay between case detection and death. FALSE corresponds
to a naive severity being calculated, TRUE corresponds to the user
calculating a corrected severity}

\item{max_date}{A string representing a user supplied maximum date, up to
which the time-varying severity estimate will be calculated. Useful in the
case of long time-series, where the user wishes to focus on a specific
time-period}
}
\value{
A data.frame containing the MLE estimate and 95\% confidence interval
of the corrected severity, named "ascertainment_me" (for the central estimate
), and "ascertainment_lo" and "ascertainment_hi" for the lower and upper
interval limits.
}
\description{
Estimates the proportion of cases or infections that have been
ascertained (declared as official cases), given a time-series of cases and
deaths, a delay distribution and a baseline severity estimate. The resulting
ascertainment estimate is calculated as the ratio of the baseline severity
estimate and the delay-adjusted severity estimate
}
\examples{
# get data pre-loaded with {incidence2}
df_covid_uk <- incidence2::covidregionaldataUK

# aggregate the covid data for the UK as a whole
df_covid_uk <- aggregate(
  data = df_covid_uk,
  x = cbind(cases_new, deaths_new) ~ date,
  FUN = function(x) sum(x, na.rm = TRUE)
)

# rename columns
colnames(df_covid_uk)[colnames(df_covid_uk) == "cases_new"] <- "cases"
colnames(df_covid_uk)[colnames(df_covid_uk) == "deaths_new"] <- "deaths"

df_covid_uk_subset <- subset(df_covid_uk, date <= "2020-05-31")

# load epidist object from {epiparameter}
onset_to_death_covid <- epiparameter::epidist_db(
  disease = "COVID-19",
  epi_dist = "onset_to_death",
  author = "Linton_etal"
)

estimate_ascertainment(
  data = df_covid_uk,
  epidist = onset_to_death_covid,
  type = "varying",
  severity_baseline = 0.014,
  smooth_inputs = TRUE,
  burn_in_value = 7L,
  correct_for_delays = TRUE,
  max_date = "2020-06-30"
)

}
