% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_ccfr.R
\name{estimate_ccfr}
\alias{estimate_ccfr}
\title{Estimate the corrected case fatality rate}
\usage{
estimate_ccfr(total_cases, total_deaths, u_t, poisson_threshold = 200)
}
\arguments{
\item{total_cases}{The total number of cases observed over the period of an
outbreak of interest. The total number of cases must be greater than or equal
to the total number of deaths}

\item{total_deaths}{The total number of deaths observed over the period of an
outbreak of interest. The total number of deaths must be less than or equal
to the total number of cases}

\item{u_t}{The proportion of cases to cases with known outcomes up to the
point of an outbreak of interest. Used to correct the total number of deaths
for delays between case detection and outcome. Given that its a proportion,
it must be between 0.0 and 1.0.}

\item{poisson_threshold}{The case count above which to use Poisson
approximation. Set to 200 by default.}
}
\value{
A named vector with the MLE and 95\% confidence interval of the
corrected CFR estimates, named "cfr_me", "cfr_low", and "cfr_high".
}
\description{
Estimate the maximum likelihood estimate and 95\% confidence
interval of a corrected CFR, using the total cases and total cases with known
outcomes, where the latter replaces the total number of deaths in the
standard (naive) CFR definition. We use a binomial likelihood, approximated
by a Poisson likelihood for large samples
}
\examples{
estimate_ccfr(total_cases = 1000, total_deaths = 900, u_t = 0.2)
}
