% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_rolling.R
\name{estimate_rolling}
\alias{estimate_rolling}
\title{Estimate static severity for an expanding time series}
\usage{
estimate_rolling(
  data,
  correct_for_delays = FALSE,
  epidist = NULL,
  poisson_threshold = 100
)
}
\arguments{
\item{data}{A data.frame containing the outbreak data. A daily time series
with dates or some other absolute indicator of time (e.g. epiday/epiweek) and
the numbers of new cases and new deaths at each time point.
Note that the required columns are "date" (for the date), "cases" (for the
number of reported cases), and "deaths" (for the number of reported deaths)
on each day of the outbreak.
Note that the data.frame is required to have an unbroken sequence of dates
with no missing dates in between. The "date" column must be of the class
\code{Date} (see \code{\link[=as.Date]{as.Date()}}).
Note also that the total number of cases must be greater than the total
number of reported deaths.}

\item{correct_for_delays}{A boolean flag indicating whether the user wishes
to correct for the delay between case detection and death. FALSE corresponds
to a naive severity being calculated, TRUE corresponds to the user
calculating a corrected severity}

\item{epidist}{The delay distribution used, in the form of an
\code{\link[epiparameter:epidist]{epiparameter::epidist()}} object. This is used to obtain a probability
mass function parameterised by time; i.e. \eqn{f(t)} which gives the
probability a case has a known outcomes (i.e. death) at time \eqn{t},
parameterised with disease-specific parameters before it is supplied here.
A typical example would be a symptom onset to death delay distribution.}

\item{poisson_threshold}{The case count above which to use Poisson
approximation. Set to 200 by default.}
}
\value{
A data.frame with the MLE and 95\% confidence interval of the
daily severity estimates, named "severity_me", "severity_lo", and
"severity_hi", with one row for each day in the original data.frame.
}
\description{
Calculates the CFR at each time point in the case and death time
series supplied, using an expanding window of time. The static CFR is
calculated for each time point, using the time series from the start to each
time point, and increasing the number of time points included by one in each
iteration.
}
\details{
When \code{correct_for_delays} is \code{TRUE}, the internal function
\code{\link[=estimate_severity]{estimate_severity()}} is used to calculate the rolling severity.
}
\examples{
# load package data
data("ebola1976")

# get an onset to death distribution from the {epiparameter} package
onset_to_death_ebola <- epiparameter::epidist_db(
  disease = "Ebola Virus Disease",
  epi_dist = "onset_to_death",
  author = "Barry_etal",
  single_epidist = TRUE
)

# estimate severity without correcting for delays
estimate_static(
  ebola1976,
  correct_for_delays = FALSE
)

# estimate severity for each day while correcting for delays
# view only the first values
head(
  estimate_rolling(
    ebola1976,
    correct_for_delays = TRUE,
    epidist = onset_to_death_ebola
  )
)

}
