% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_severity.R
\name{estimate_severity}
\alias{estimate_severity}
\title{Estimate the corrected case fatality rate}
\usage{
estimate_severity(df_in, poisson_threshold = 100, location)
}
\arguments{
\item{df_in}{A data.frame containing the outbreak data. A daily time series
with dates or some other absolute indicator of time (e.g. epiday/epiweek) and
the numbers of new cases and new deaths at each time point}

\item{poisson_threshold}{The case count above which to use Poisson
approximation. Set to 200 by default.}

\item{location}{Column name as a string by which to disaggregate data for
estimates. This is usually a column indicating the region or country to which
the estimate applies.}
}
\value{
A data.frame with the MLE and 95\% confidence interval of the
corrected severity estimates, named "severity_me", "severity_low", and
"severity_high".
}
\description{
Estimate the maximum likelihood estimate and 95\% confidence
interval of a corrected severity, using the total cases and total cases with
known outcomes, where the latter replaces the total number of deaths in the
standard (naive) severity definition. We use a binomial likelihood,
approximated by a Poisson likelihood for large samples
}
\keyword{internal}
