% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_static.R
\name{estimate_static}
\alias{estimate_static}
\title{Estimate a static (in time) severity measure}
\usage{
estimate_static(
  data,
  epidist,
  correct_for_delays = TRUE,
  poisson_threshold = 100
)
}
\arguments{
\item{data}{A data.frame containing the outbreak data. A daily time series
with dates or some other absolute indicator of time (e.g. epiday/epiweek) and
the numbers of new cases and new deaths at each time point.
Note that the required columns are "date" (for the date), "cases" (for the
number of reported cases), and "deaths" (for the number of reported deaths)
on each day of the outbreak.
Note that the data.frame is required to have an unbroken sequence of dates
with no missing dates in between. The "date" column must be of the class
\code{Date} (see \code{\link[=as.Date]{as.Date()}}).
Note also that the total number of cases must be greater than the total
number of reported deaths.}

\item{epidist}{The delay distribution used, in the form of an
\code{\link[epiparameter:epidist]{epiparameter::epidist()}} object. This is used to obtain a probability
mass function parameterised by time; i.e. \eqn{f(t)} which gives the
probability a case has a known outcomes (i.e. death) at time \eqn{t},
parameterised with disease-specific parameters before it is supplied here.
A typical example would be a symptom onset to death delay distribution.}

\item{correct_for_delays}{A boolean flag indicating whether the user wishes
to correct for the delay between case detection and death. FALSE corresponds
to a naive severity being calculated, TRUE corresponds to the user
calculating a corrected severity}

\item{poisson_threshold}{The case count above which to use Poisson
approximation. Set to 100 by default.}
}
\value{
A data.frame with the MLE and 95\% confidence interval of the
severity estimates, named "severity_mean", "severity_low", and
"severity_high".
}
\description{
Calculates the severity of a disease, corrected for a
user-specified delay. If cases are supplied, and the delay distribution
representing the delay between case detection and death, then a case
fatality ratio over time is estimated
}
\examples{
# load package data
data("ebola1976")

# get an onset to death distribution from the {epiparameter} package
onset_to_death_ebola <- epiparameter::epidist_db(
  disease = "Ebola Virus Disease",
  epi_dist = "onset_to_death",
  author = "Barry_etal",
  single_epidist = TRUE
)

# estimate severity without correcting for delays
estimate_static(
  ebola1976,
  correct_for_delays = FALSE
)

# estimate severity while correcting for delays
estimate_static(
  ebola1976,
  correct_for_delays = TRUE,
  epidist = onset_to_death_ebola
)

}
