% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_cfr_neatly.R
\name{format_cfr_neatly}
\alias{format_cfr_neatly}
\title{Neatly format CFR estimates}
\usage{
format_cfr_neatly(cfr_in, type)
}
\arguments{
\item{cfr_in}{A named vector of the format outputted by \code{\link[=estimate_ccfr]{estimate_ccfr()}}
or \code{\link[=static_cfr]{static_cfr()}}}

\item{type}{A string describing the type of severity estimate. Typical
entries would be CFR, HFR, Naive, Corrected, etc.}
}
\value{
Prints the output of \code{\link[=estimate_ccfr]{estimate_ccfr()}} or \code{\link[=static_cfr]{static_cfr()}} neatly to
the R terminal
}
\description{
Format the CFR MLE and 95\% range estimates neatly, giving the
estimates to 3 significant figures and presenting the MLE and 95\% interval
on a single line
}
\examples{
data("ebola1976")

df_ebola_subset <- subset(ebola1976, date <= "1976-09-18")

onset_to_death_ebola <- epiparameter::epidist_db(
  disease = "Ebola Virus Disease", 
  epi_dist = "onset_to_death",
  author = "Barry_etal")
  
static_cfr(
  df_ebola_subset,
  correct_for_delays = FALSE) |> 
  format_cfr_neatly(type = "Naive")
  
}
