% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_output.R
\name{format_output}
\alias{format_output}
\title{Neatly format CFR estimates}
\usage{
format_output(data, estimate_type, type = NULL)
}
\arguments{
\item{data}{A data.frame of the format outputted by \code{\link[=estimate_static]{estimate_static()}}
or \code{\link[=estimate_time_varying]{estimate_time_varying()}}.}

\item{estimate_type}{A required string describing whether a severity of
reporting estimate is to be formatted. The package calculates both severity
and under-ascertainment estimates and uses the same formatting function for
both}

\item{type}{An optional string describing the estimate, typically used to
describe whether the estimate has been corrected for delays or not}
}
\value{
Prints the output of \code{\link[=estimate_static]{estimate_static()}} or \code{\link[=estimate_time_varying]{estimate_time_varying()}}
neatly to the R terminal.
}
\description{
Format the CFR MLE and 95\% range estimates neatly, giving the
estimates to 3 significant figures and presenting the MLE and 95\% interval
on a single line
}
\examples{
# load Ebola 1976 outbreak data
data("ebola1976")

# add the location the dataset relates to, for ease of formatting at the end
# of this example
ebola1976$location <- "Democratic Republic of the Congo"

# subset the data so we focus on the earlier half of the outbreak
df_ebola_subset <- subset(ebola1976, date <= "1976-09-30")

# read epidist for EVD onset to death from {epiparameter}
onset_to_death_ebola <- epiparameter::epidist_db(
  disease = "Ebola Virus Disease",
  epi_dist = "onset_to_death",
  author = "Barry_etal"
)

# Calculate static naive CFR
ncfr <- estimate_static(
  data = df_ebola_subset,
  correct_for_delays = FALSE
)

# Calculate static corrected CFRs
ccfr <- estimate_static(
  data = ebola1976,
  correct_for_delays = TRUE,
  epi_dist = onset_to_death_ebola
)

# Formats the output of the CFR data.frames nicely
# and prints to the terminal
format_output(ncfr, estimate_type = "severity", type = "Naive")
format_output(ccfr, estimate_type = "severity", type = "Corrected")

}
