% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_time_varying.R
\name{get_default_burn_in}
\alias{get_default_burn_in}
\title{Get a default burn-in value from a delay distribution}
\usage{
get_default_burn_in(epidist)
}
\arguments{
\item{epidist}{An optional argument for the delay distribution used, in the
form of an \verb{<epidist>} object; see \code{\link[epiparameter:epidist]{epiparameter::epidist()}}.

Passing an \verb{<epidist>} automatically applies delay correction in the
severity estimation.}
}
\value{
A single integer, the burn-in value.
}
\description{
Get a default burn-in value from a delay distribution
}
\keyword{internal}
