% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_density}
\alias{get_density}
\title{Get the PDF/PMF from a distribution}
\usage{
get_density(f, x)
}
\arguments{
\item{f}{An object specifying a distribution, or a closure wrapping a call to
the density function on a distribution object.
Only \verb{<epidist>} objects from \pkg{epiparameter} and closures are currently
supported. See \strong{Examples}.}

\item{x}{A sequence of numeric values at which to estimate the density.}
}
\value{
A numeric vector for the distribution density at the values passed in \code{x}.
}
\description{
Calculates the density of the distribution at user-specified
values. PDF and PMF apply for continuous and discrete distributions,
respectively.
}
\keyword{internal}
