% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_case_data.R
\name{plot_case_data}
\alias{plot_case_data}
\title{Plot the raw case time-series data}
\usage{
plot_case_data(data)
}
\arguments{
\item{data}{A data.frame containing the outbreak data. A daily time series
with dates or some other absolute indicator of time (e.g. epiday/epiweek) and
the numbers of new case at each time point. The input data.frame assumes
columns with the names: date, cases and deaths}
}
\value{
A plot of the case time-series
}
\description{
Produces a simple plot, using base R, of the raw case
time-series data
}
\examples{
# Load Ebola 1976 outbreak data
data("ebola1976")

plot_case_data(ebola1976)
}
