% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_death_data.R
\name{plot_death_data}
\alias{plot_death_data}
\title{Plot the raw death time-series data}
\usage{
plot_death_data(df_in)
}
\arguments{
\item{df_in}{A data.frame containing the outbreak data. A daily time series
with dates or some other absolute indicator of time (e.g. epiday/epiweek) and
the numbers of new deaths at each time point. The input data.frame assumes
columns with the names: date, cases and deaths}
}
\value{
A plot of the death time-series
}
\description{
Produces a simple plot, using base R, of the raw death
time-series data
}
\examples{
# Load Ebola 1976 outbreak data
data("ebola1976")

plot_death_data(ebola_1976)
}
