% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_known_outcomes.R
\name{plot_known_outcomes}
\alias{plot_known_outcomes}
\title{Plot the estimated known outcomes, as well as the raw data}
\usage{
plot_known_outcomes(df_in, together = TRUE)
}
\arguments{
\item{df_in}{A data.frame of the format that the known_outcomes() function returns from the
datadelay package}
}
\value{
A plot of the three time-series on the same plot, with a legend
}
\description{
Produces a simple plot, using base R, of the raw case and death time-series data,
as well as the estimated number of known outcomes on each day
}
\examples{
# Load Ebola 1976 outbreak data
data("ebola1976")

onset_to_death_ebola <- epiparameter::epidist_db(
disease = "Ebola Virus Disease",
epi_dist = "onset_to_death",
author = "Barry_etal")

df_known_outcomes <- known_outcomes(
df_in = ebola1976,
epi_dist = onset_to_death_ebola)

plot_known_outcomes(df_known_outcomes)
}
