% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_raw_data.R
\name{plot_raw_data}
\alias{plot_raw_data}
\title{Plot the raw case and death time-series data}
\usage{
plot_raw_data(df_in, together = TRUE)
}
\arguments{
\item{df_in}{A data.frame containing the outbreak data. A daily time series
with dates or some other absolute indicator of time (e.g. epiday/epiweek) and
the numbers of new cases and new deaths at each time point}

\item{together}{A boolean flag which determines whether both the
cases/known outcomes are plotted in the same panel as the deaths. If
the two are on significantly different scales, plot them separately.}
}
\value{
A plot of the two time-series on the same plot, with a legend
}
\description{
Produces a simple plot, using base R, of the raw case and death
time-series data
}
\examples{
# Load Ebola 1976 outbreak data
data("ebola1976")

plot_raw_data(ebola1976)
}
