% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_time_varying.R
\name{plot_time_varying}
\alias{plot_time_varying}
\title{Plot time-varying severity estimates}
\usage{
plot_time_varying(data, lower = 0, upper = 10)
}
\arguments{
\item{data}{A data.frame of the format returned by \code{\link[=known_outcomes]{known_outcomes()}}.}

\item{lower}{A numeric value determining the lower limit of the y-axis of the
plot. In place to easily help the user produce a readable graph, as severity
estimates often have typical ranges for certain diseases.}

\item{upper}{A numeric value determining the upper limit of the y-axis of the
plot.}
}
\value{
A plot of the time-varying severity estimate
}
\description{
Produces a plot, using base R, of the time-varying severity
estimate over the entire period it was calculated.
}
\examples{
library(datadelay)
library(epiparameter)
library(covidregionaldata)

df_covid_uk <- get_national_data(
  countries = "united kingdom", source = "who", verbose = FALSE
)
# rename columns
colnames(df_covid_uk)[colnames(df_covid_uk) == "cases_new"] <- "cases"
colnames(df_covid_uk)[colnames(df_covid_uk) == "deaths_new"] <- "deaths"

df_covid_uk_subset <- subset(df_covid_uk, date <= "2020-12-31")

onset_to_death_covid <- epidist_db(
  disease = "COVID-19",
  epi_dist = "onset_to_death",
  author = "Linton_etal"
)

df_covid_cfr_uk_naive <- estimate_time_varying(
  df_covid_uk_subset,
  epi_dist = onset_to_death_covid,
  smooth_inputs = TRUE,
  burn_in_value = 7,
  correct_for_delays = FALSE
)

df_covid_cfr_uk_corrected <- estimate_time_varying(
  df_covid_uk_subset,
  epi_dist = onset_to_death_covid,
  smooth_inputs = TRUE,
  burn_in_value = 7,
  correct_for_delays = TRUE
)

plot_time_varying(df_covid_cfr_uk_naive, lower = 0, upper = 5)
plot_time_varying(df_covid_cfr_uk_corrected, lower = 0, upper = 5)

}
