% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data.R
\name{prepare_data.incidence2}
\alias{prepare_data.incidence2}
\title{Prepare data from \if{html}{\out{<incidence2>}} objects}
\usage{
\method{prepare_data}{incidence2}(data, cases_variable = "cases", deaths_variable = "deaths")
}
\description{
Prepare data from \if{html}{\out{<incidence2>}} objects
}
