% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data.R
\name{prepare_data}
\alias{prepare_data}
\title{Prepare data for CFR estimation}
\usage{
prepare_data(data, ...)
}
\arguments{
\item{data}{A \code{data.frame}-like object.}

\item{...}{Other arguments passed to methods.}
}
\description{
This S3 generic has no default method. Rather, it has methods for classes
commonly found in epidemiological data, such as \verb{<incidence2>} from the
\code{{incidence2}} package. See \code{\link[incidence2:incidence]{incidence2::incidence()}}.
}
