% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/static_cfr.R
\name{static_cfr}
\alias{static_cfr}
\title{Calculates the CFR at a single point in time using in the case and death time
series supplied. Uses all of the time points supplied by the user and returns
an estimate for a single point in time}
\usage{
static_cfr(df_in, correct_for_delays = TRUE, delay_pmf)
}
\arguments{
\item{df_in}{A data.frame containing the outbreak data. A daily time series
with dates or some other absolute indicator of time (e.g. epiday/epiweek) and
the numbers of new cases and new deaths at each time point}

\item{correct_for_delays}{A boolean flag indicating whether the user wishes
to correct for the delay between case detection and death. FALSE corresponds
to a naive CFR being calculated, TRUE corresponds to the user calculating a
corrected CFR}

\item{delay_pmf}{The delay distribution used, in the form of a probability
mass function parameterised by time. I.e. f(t) which gives the probability a
case has a known outcomes (i.e. death) at time t, parameterised with
disease-specific parameters before it is supplied here. A typical example
would be a symptom onset to death delay distribution}
}
\value{
A data.frame containing the MLE and 95\% confidence interval of the
CFR estimates
}
\description{
Calculates the CFR at a single point in time using in the case and death time
series supplied. Uses all of the time points supplied by the user and returns
an estimate for a single point in time
}
\examples{
library(lubridate)
library(epiparameter)

# read in the dataset in question. 
# dataset needs 
df_ebola <- read.csv("data/ebola_1976.csv")

df_ebola$date <- ymd(df_ebola$date)

# read in onset to death distribution
onset_to_death_ebola <- epidist("ebola","onset_to_death")$pmf

calculate static naive and corrected CFRs
ncfr <- calcalate_static_cfr(df_ebola, correct_for_delays = FALSE)
ccfr <- calcalate_static_cfr(df_ebola, correct_for_delays = TRUE, onset_to_death_ebola)

# just a function that formats the output of the CFR data.frames 
# nicely and prints to the terminal
format_cfr_output_neatly(ncfr)
format_cfr_output_neatly(ccfr)
}
