% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_delay_density.R
\name{test_fn_req_args}
\alias{test_fn_req_args}
\title{Check delay density functions passed to \verb{cfr_*()}}
\usage{
test_fn_req_args(fn, n_req_args = 1)
}
\arguments{
\item{fn}{A function.}

\item{n_req_args}{The number of required arguments, i.e., arguments without
default values.}
}
\value{
A logical for whether the function \code{fn} has the number of required
arguments specified by the user.
}
\description{
Checks whether a function has a given number of required arguments.
}
\keyword{internal}
