% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cases_to_infections.R
\name{cases_to_infections}
\alias{cases_to_infections}
\title{Cases to infections}
\usage{
cases_to_infections(
  case_times,
  case_data,
  infection_to_onset,
  onset_to_report,
  pre_window = 20
)
}
\arguments{
\item{case_times}{Time values corresponding to case incidence}

\item{case_data}{Time series of case incidence data}

\item{infection_to_onset}{Probability mass function for infection-to-onset}

\item{onset_to_report}{Probability mass function for onset-to-report}

\item{pre_window}{Period to reconstruct before first case data point}
}
\description{
Function to estimate infection incidence from case data.
}
\examples{
man/examples/cases_to_infections_example
cases_to_infections()
}
\keyword{cases}
