% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_rolling.R
\name{cfr_rolling}
\alias{cfr_rolling}
\title{Estimate static severity for an expanding time series}
\usage{
cfr_rolling(data, epidist = NULL, poisson_threshold = 100)
}
\arguments{
\item{data}{A \verb{<data.frame>} containing the outbreak data. A daily time
series with dates or some other absolute indicator of time (e.g. epiday or
epiweek) and the numbers of new cases and new deaths at each time point.
Note that the required columns are "date" (for the date), "cases" (for the
number of reported cases), and "deaths" (for the number of reported deaths)
on each day of the outbreak.

Note that the \verb{<data.frame>} is required to have an unbroken sequence of
dates with no missing dates in between. The "date" column must be of class
\code{Date} (see \code{\link[=as.Date]{as.Date()}}).

Note also that the total number of cases must be greater than the total
number of reported deaths.}

\item{epidist}{An optional argument for the delay distribution used, in the
form of an \verb{<epidist>} object.
Defaults to \code{NULL}, in which case no delay correction is applied.
Passing an \verb{<epidist>} automatically applies delay correction in the
severity estimation.}

\item{poisson_threshold}{The case count above which to use Poisson
approximation. Set to 100 by default.}
}
\value{
A \verb{<data.frame>} with the MLE and 95\% confidence interval of the
daily severity estimates, named "severity_mean", "severity_low", and
"severity_high", with one row for each day in the original data.frame.
}
\description{
Calculates the CFR at each time point in the case and death time
series supplied, using an expanding window of time. The static CFR is
calculated for each time point, using the time series from the start to each
time point, and increasing the number of time points included by one in each
iteration.
}
\details{
When delay correction is applied by passing an \verb{<epidist>}, the
internal function \code{\link[=estimate_severity]{estimate_severity()}} is used to calculate the rolling
severity.
}
\examples{
# load package data
data("ebola1976")

# get an onset to death distribution from the {epiparameter} package
onset_to_death_ebola <- epiparameter::epidist_db(
  disease = "Ebola Virus Disease",
  epi_dist = "onset_to_death",
  author = "Barry_etal",
  single_epidist = TRUE
)

# estimate severity without correcting for delays
cfr_static(ebola1976)

# estimate severity for each day while correcting for delays
# view only the first values
estimate <- cfr_rolling(
  ebola1976,
  epidist = onset_to_death_ebola
)

head(estimate)

}
