% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadelay-package.R
\docType{package}
\name{datadelay-package}
\alias{datadelay}
\alias{datadelay-package}
\title{datadelay: Delay Functions for Epidemic Data}
\description{
Delay functions for routine surveillance data.
}
\author{
\strong{Maintainer}: Adam Kucharski \email{adam.kucharski@lshtm.ac.uk} (\href{https://orcid.org/0000-0001-8814-9421}{ORCID})

Authors:
\itemize{
  \item Tim Russell \email{timothy.russell@lshtm.ac.uk} (\href{https://orcid.org/0000-0001-5610-6080}{ORCID})
}

}
\keyword{internal}
