% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebola1976.R
\docType{data}
\name{ebola1976}
\alias{ebola1976}
\title{Ebola 1976 outbreak case data}
\format{
\subsection{\code{ebola1976}}{

A data frame with 73 rows and 3 columns:
\describe{
\item{date}{Calendar date}
\item{cases}{Number of cases reported}
\item{deaths}{Number of deaths reported}
}
}
}
\source{
\doi{10.1016/j.epidem.2014.09.003}
}
\usage{
ebola1976
}
\description{
An example epidemic outbreak dataset for use with the \code{cfr} package.
This dataset comes from the first Ebola outbreak in Zaire in 1976 as analysed
in Camacho et al. (2014).
}
\keyword{datasets}
