% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_severity.R
\name{estimate_severity}
\alias{estimate_severity}
\title{Estimate the corrected case fatality rate}
\usage{
estimate_severity(data, poisson_threshold = 100)
}
\arguments{
\item{data}{A data.frame containing the outbreak data. A daily time series
with dates or some other absolute indicator of time (e.g. epiday/epiweek) and
the numbers of new cases and new deaths at each time point.
Note that the required columns are "date" (for the date), "cases" (for the
number of reported cases), and "deaths" (for the number of reported deaths)
on each day of the outbreak.
Note that the data.frame is required to have an unbroken sequence of dates
with no missing dates in between. The "date" column must be of the class
\code{Date} (see \code{\link[=as.Date]{as.Date()}}).
Note also that the total number of cases must be greater than the total
number of reported deaths.}

\item{poisson_threshold}{The case count above which to use Poisson
approximation. Set to 200 by default.}
}
\value{
A data.frame with the MLE and 95\% confidence interval of the
corrected severity estimates, named "severity_me", "severity_low", and
"severity_high".
}
\description{
Estimate the maximum likelihood estimate and 95\% confidence
interval of a corrected severity, using the total cases and total cases with
known outcomes, where the latter replaces the total number of deaths in the
standard (naive) severity definition. We use a binomial likelihood,
approximated by a Poisson likelihood for large samples
}
\keyword{internal}
