% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_severity.R
\name{estimate_severity}
\alias{estimate_severity}
\title{Estimate the corrected case fatality risk}
\usage{
estimate_severity(
  total_cases,
  total_deaths,
  total_outcomes,
  poisson_threshold = 100
)
}
\arguments{
\item{total_cases}{The total number of cases observed over the period of an
outbreak of interest. The total number of cases must be greater than or equal
to the total number of deaths.}

\item{total_deaths}{The total number of deaths observed over the period of an
outbreak of interest. The total number of deaths must be less than or equal
to the total number of cases.}

\item{total_outcomes}{The total number of outcomes expected to be observed
over the period of an outbreak of interest. See \code{\link[=known_outcomes]{known_outcomes()}}.}

\item{poisson_threshold}{The case count above which to use Poisson
approximation. Set to 100 by default.}
}
\value{
A data.frame with the maximum likelihood estimate and 95\% confidence
interval of the corrected severity estimates, named "severity_mean",
"severity_low", and "severity_high".
}
\description{
Estimates the maximum likelihood estimate and 95\% confidence
interval of a corrected severity, using the total cases and total cases with
known outcomes, where the latter replaces the total number of deaths in the
standard (naive) severity definition. We use a binomial likelihood,
approximated by a Poisson likelihood for large samples.
}
\keyword{internal}
