% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_severity.R
\name{estimate_severity}
\alias{estimate_severity}
\title{Estimate the corrected case fatality rate}
\usage{
estimate_severity(df_in, poisson_threshold = 100, group_by = "country")
}
\arguments{
\item{total_cases}{The total number of cases observed over the period of an
outbreak of interest. The total number of cases must be greater than or equal
to the total number of deaths}

\item{total_deaths}{The total number of deaths observed over the period of an
outbreak of interest. The total number of deaths must be less than or equal
to the total number of cases}

\item{u_t}{The proportion of cases to cases with known outcomes up to the
point of an outbreak of interest. Used to correct the total number of deaths
for delays between case detection and outcome. Given that it is a proportion,
it must be between 0.0 and 1.0.}
}
\value{
A named vector with the MLE and 95\% confidence interval of the
corrected severity estimates, named "severity_me", "severity_low", and
"severity_high".
}
\description{
Estimate the maximum likelihood estimate and 95\% confidence
interval of a corrected severity, using the total cases and total cases with
known outcomes, where the latter replaces the total number of deaths in the
standard (naive) severity definition. We use a binomial likelihood,
approximated by a Poisson likelihood for large samples
}
\keyword{internal}
