% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_static.R
\name{estimate_static}
\alias{estimate_static}
\title{Estimate a static (in time) severity measure}
\usage{
estimate_static(
  data,
  correct_for_delays = TRUE,
  epi_dist,
  poisson_threshold = 100
)
}
\arguments{
\item{data}{A data.frame containing the outbreak data. A daily time series
with dates or some other absolute indicator of time (e.g. epiday/epiweek) and
the numbers of new cases and new deaths at each time point.
Note that the required columns are "date" (for the date), "cases" (for the
number of reported cases), and "deaths" (for the number of reported deaths)
on each day of the outbreak.
Note that the data.frame is required to have an unbroken sequence of dates
with no missing dates in between. The "date" column must be of the class
\code{Date} (see \code{\link[=as.Date]{as.Date()}}).
Note also that the total number of cases must be greater than the total
number of reported deaths.}

\item{correct_for_delays}{A boolean flag indicating whether the user wishes
to correct for the delay between case detection and death. FALSE corresponds
to a naive severity being calculated, TRUE corresponds to the user
calculating a corrected severity}

\item{epi_dist}{The delay distribution used, in the form of an
\code{\link[epiparameter:epidist]{epiparameter::epidist()}} object. This is used to obtain a probability
mass function parameterised by time; i.e. \eqn{f(t)} which gives the
probability a case has a known outcomes (i.e. death) at time \eqn{t},
parameterised with disease-specific parameters before it is supplied here.
A typical example would be a symptom onset to death delay distribution.}

\item{poisson_threshold}{The case count above which to use Poisson
approximation. Set to 200 by default.}
}
\value{
A data.frame with the MLE and 95\% confidence interval of the
severity estimates, named "severity_me", "severity_low", and "severity_high",
as well as a column "location" holding the grouping variable indicating the
region.
}
\description{
Calculates the severity of a disease, corrected for a
user-specified delay. If cases are supplied, and the delay distribution
representing the delay between case detection and death, then a case
fatality ratio over time is estimated
}
\examples{
library(datadelay)
library(epiparameter)
library(covidregionaldata)

data("ebola1976")

ebola1976$location <- "Democractic Republic of the Congo"

df_ebola_subset <- subset(ebola1976, date <= "1976-09-30")

onset_to_death_ebola <- epidist_db(
  disease = "Ebola Virus Disease",
  epi_dist = "onset_to_death",
  author = "Barry_etal"
)

df_ncfr_static_ebola <- estimate_static(
  df_ebola_subset,
  correct_for_delays = FALSE
)

format_output(
  df_ncfr_static_ebola,
  estimate_type = "severity",
  type = "Naive CFR"
)

# calculating the corrected CFR
df_ccfr_static_ebola <- estimate_static(
  df_ebola_subset,
  correct_for_delays = TRUE,
  epi_dist = onset_to_death_ebola
)

format_output(
  df_ccfr_static_ebola,
  estimate_type = "severity",
  type = "Corrected CFR"
)

}
