% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_cfr_neatly.R
\name{format_cfr_neatly}
\alias{format_cfr_neatly}
\title{Neatly format CFR estimates}
\usage{
format_cfr_neatly(cfr_in)
}
\arguments{
\item{cfr_in}{A named vector of the format outputted by \code{\link[=estimate_ccfr]{estimate_ccfr()}}
or \code{\link[=static_cfr]{static_cfr()}}}
}
\value{
Prints the output of \code{\link[=estimate_ccfr]{estimate_ccfr()}} or \code{\link[=static_cfr]{static_cfr()}} neatly to
the R terminal
}
\description{
Format the CFR MLE and 95\% range estimates neatly, giving the
estimates to 3 significant figures and presenting the MLE and 95\% interval
on a single line
}
\examples{
# Create an `epidist` for EVD onset to death distribution
# taken from parameters in 10.1016/S0140-6736(18)31387-4
onset_to_death_ebola <- epiparameter::epidist(
  disease = "Ebola virus disease",
  pathogen = "Ebolavirus",
  epi_dist = "onset_to_death",
  prob_distribution = "gamma",
  prob_distribution_params = c(
    shape = 2.4, scale = 3.333
  )
)

# load Ebola 1976 outbreak data
data("ebola1976")
# Calculate static naive CFR
ncfr <- static_cfr(df_in = ebola1976, correct_for_delays = FALSE)

# Calculate static corrected CFRs
ccfr <- static_cfr(
  df_in = ebola1976,
  correct_for_delays = TRUE,
  epi_dist = onset_to_death_ebola
)

# Formats the output of the CFR data.frames nicely
# and prints to the terminal
format_cfr_neatly(ncfr)
format_cfr_neatly(ccfr)
}
