% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_cfr_neatly.R
\name{format_cfr_neatly}
\alias{format_cfr_neatly}
\title{Function to format the CFR MLE and 95\% range estimates neatly, giving the
estimates to 3 significant figures and presenting the MLE and 95\% interval
on a single line}
\usage{
format_cfr_neatly(df_in)
}
\arguments{
\item{df_in}{A data.frame of the format outputted by ccfr_uncertainty.R}
}
\value{
Prints the output of ccfr_uncertainty.R neatly to the R terminal
}
\description{
Function to format the CFR MLE and 95\% range estimates neatly, giving the
estimates to 3 significant figures and presenting the MLE and 95\% interval
on a single line
}
\examples{
format_cfr_neatly(data.frame(cfr_me = 0.5, cfr_lo = 0.2, cfr_hi = 0.7))
}
