% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_cfr_neatly.R
\name{format_cfr_neatly}
\alias{format_cfr_neatly}
\title{Neatly format CFR estimates}
\usage{
format_cfr_neatly(cfr_in, type)
}
\arguments{
\item{cfr_in}{A named vector of the format outputted by \code{\link[=estimate_ccfr]{estimate_ccfr()}}
or \code{\link[=static_cfr]{static_cfr()}}}
}
\value{
Prints the output of \code{\link[=estimate_ccfr]{estimate_ccfr()}} or \code{\link[=static_cfr]{static_cfr()}} neatly to
the R terminal
}
\description{
Format the CFR MLE and 95\% range estimates neatly, giving the
estimates to 3 significant figures and presenting the MLE and 95\% interval
on a single line
}
\examples{
# read epidist for EVD onset to death from {epiparameter}
onset_to_death_ebola <- epiparameter::epidist_db(
  disease = "Ebola Virus Disease",
  epi_dist = "onset_to_death",
  author = "Barry_etal"
)

# load Ebola 1976 outbreak data
data("ebola1976")
# Calculate static naive CFR
ncfr <- static_cfr(df_in = ebola1976, correct_for_delays = FALSE)

# Calculate static corrected CFRs
ccfr <- static_cfr(
  df_in = ebola1976,
  correct_for_delays = TRUE,
  epi_dist = onset_to_death_ebola
)

# Formats the output of the CFR data.frames nicely
# and prints to the terminal
format_cfr_neatly(ncfr)
format_cfr_neatly(ccfr)
}
