% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_time_varying.R
\name{get_default_burn_in}
\alias{get_default_burn_in}
\title{Get a default burn-in value from a delay distribution}
\usage{
get_default_burn_in(delay_dist = NULL)
}
\arguments{
\item{delay_dist}{An optional argument that specifies how the duration
between cases and outcomes is distributed.
May be one of two types:
\enumerate{
\item A delay distribution in the form of an \verb{<epidist>} object, the density of
which is evaluated by calling the \code{stats::density()} generic, or,
\item A closure that wraps the density function of a distribution to evaluate
density at user-specified values, e.g.
\code{function(x) stats::dgamma(shape = 5, scale = 1, x = x)}
}

Defaults to \code{NULL}, in which case no delay correction is applied.
Passing an \verb{<epidist>} or closure automatically applies delay correction in
the severity estimation.}
}
\value{
A single integer, the burn-in value.
}
\description{
Get a default burn-in value from a delay distribution
}
\keyword{internal}
