% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_time_varying.R
\name{get_default_burn_in}
\alias{get_default_burn_in}
\title{Get a default burn-in value from a delay distribution}
\usage{
get_default_burn_in(epi_dist = NULL)
}
\arguments{
\item{epi_dist}{The delay distribution used, in the form of an
\code{\link[epiparameter:epidist]{epiparameter::epidist()}} object. This is used to obtain a probability
mass function parameterised by time; i.e. \eqn{f(t)} which gives the
probability a case has a known outcomes (i.e. death) at time \eqn{t},
parameterised with disease-specific parameters before it is supplied here.
A typical example would be a symptom onset to death delay distribution.}
}
\value{
A single integer, the burn-in value.
}
\description{
Get a default burn-in value from a delay distribution
}
\keyword{internal}
