% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/known_outcomes.R
\name{known_outcomes}
\alias{known_outcomes}
\title{Estimate known outcomes from case and death time-series data}
\usage{
known_outcomes(df_in, epi_dist, cumulative = TRUE)
}
\arguments{
\item{df_in}{A data.frame containing the outbreak data. A daily time series
with dates or some other absolute indicator of time (e.g. epiday/epiweek) and
the numbers of new cases and new deaths at each time point}

\item{epi_dist}{The delay distribution used, in the form of an
\code{\link[epiparameter:epidist]{epiparameter::epidist()}} object. This is used to obtain a probability
mass function parameterised by time; i.e. \eqn{f(t)} which gives the
probability a case has a known outcomes (i.e. death) at time \eqn{t},
parameterised with disease-specific parameters before it is supplied here.
A typical example would be a symptom onset to death delay distribution.}

\item{cumulative}{A boolean flag to indicate whether the user wants the daily
or total number of known outcomes}
}
\value{
A data.frame containing the MLE estimate and 95\% confidence interval
of the corrected CFR
}
\description{
Estimates the expected number of individuals with known outcomes
from a case and death time series of data of an outbreak, up to the time
point supplied.
Can either calculate the daily new number of known outcomes or the cumulative
number.
Uses the probability mass function representing the delay between
case detection and death, typically approximated by a symptom onset to death
distribution from the literature for the disease in question.
}
\examples{
# Load Ebola 1976 outbreak data
data("ebola1976")

# Create an `epidist` for EVD onset to death distribution
# taken from parameters in 10.1016/S0140-6736(18)31387-4
onset_to_death_ebola <- epiparameter::epidist(
  disease = "Ebola virus disease",
  pathogen = "Ebolavirus",
  epi_dist = "onset_to_death",
  prob_distribution = "gamma",
  prob_distribution_params = c(
    shape = 2.4, scale = 3.333
  )
)

df_known_outcomes <- known_outcomes(df_in = ebola1976, onset_to_death_ebola)
}
