% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/known_outcomes.R
\name{known_outcomes}
\alias{known_outcomes}
\title{Estimate known outcomes from case and death time-series data}
\usage{
known_outcomes(data, epi_dist)
}
\arguments{
\item{data}{A data.frame containing the outbreak data. A daily time series
with dates or some other absolute indicator of time (e.g. epiday/epiweek) and
the numbers of new cases and new deaths at each time point}

\item{epi_dist}{The delay distribution used, in the form of an
\code{\link[epiparameter:epidist]{epiparameter::epidist()}} object. This is used to obtain a probability
mass function parameterised by time; i.e. \eqn{f(t)} which gives the
probability a case has a known outcomes (i.e. death) at time \eqn{t},
parameterised with disease-specific parameters before it is supplied here.
A typical example would be a symptom onset to death delay distribution.}
}
\value{
A data.frame with the columns in \code{data}, and with two additional
columns,
\itemize{
\item "known_outcomes" for the total number of known outcomes on that day
of the outbreak, and
\item "u_t" for the under-reporting factor.
}
}
\description{
Estimates the expected number of individuals with known outcomes
from a case and death time series of data of an outbreak, up to the time
point supplied.
Calculates the daily new number of known outcomes.

Uses the probability mass function representing the delay between
case detection and death, typically approximated by a symptom onset to death
distribution from the literature for the disease in question.
}
\examples{
# Load Ebola 1976 outbreak data
data("ebola1976")

# read epidist for EVD onset to death from {epiparameter}
onset_to_death_ebola <- epiparameter::epidist_db(
  disease = "Ebola Virus Disease",
  epi_dist = "onset_to_death",
  author = "Barry_etal"
)

df_known_outcomes <- known_outcomes(data = ebola1976, onset_to_death_ebola)
}
