% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_data_and_cfr}
\alias{plot_data_and_cfr}
\title{Plot rolling case and death data with CFR estimates}
\usage{
plot_data_and_cfr(df_ncfr, df_ccfr)
}
\arguments{
\item{df_ncfr}{A data.frame containing the rolling naive CFR estimates, of
the form returned by rolling_cfr.R}

\item{df_ccfr}{A data.frame containing the rolling corrected CFR estimates,
of the form returned by rolling_cfr.R}
}
\value{
A ggplot object of the case and death time series used to calculate
the CFR estimates in one panel (top panel) and the resulting naive and
corrected CFR estimates in another panel (bottom panel)
}
\description{
Plots the case and death time series used to calculate the CFR
estimates in one panel and the resulting rolling naive and corrected CFR
estimates in another panel.
}
\examples{
# Read in onset to death distribution for Ebola
onset_to_death_ebola <- epiparameter::epidist("ebola", "onset_to_death")$pmf

# Calculate rolling naive and corrected CFRs
df_ncfr <- rolling_cfr(ebola1976, correct_for_delays = FALSE)
df_ccfr <- rolling_cfr(ebola1976,
  correct_for_delays = TRUE,
  onset_to_death_ebola
)

# Plotting rolling estimates
plot_data_and_cfr(df_ncfr, df_ccfr)

}
