% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_epiparameter_distribution.R
\name{plot_epiparameter_distribution}
\alias{plot_epiparameter_distribution}
\title{Plot an epiparameter distribution object}
\usage{
plot_epiparameter_distribution(epidist, from = 0, to = 30, by = 0.1)
}
\arguments{
\item{epidist}{An epiparameter object retrieved using the epiparameter::epidist_db()}

\item{from}{The day from which to start plotting the distribution. Default option is day 0.}

\item{to}{The day to which to start plotting the distribution. Default option is day 30.}

\item{by}{The time resolution at which which to plot the distribution. Default option is 0.1 days.}
}
\value{
A plot of the epiparameter distribution over time
}
\description{
Produces a simple plot, using base R, of an epiparameter distribution object
}
\examples{

onset_to_death_ebola <- epiparameter::epidist_db(
disease = "Ebola Virus Disease",
epi_dist = "onset_to_death",
author = "Barry_etal")

plot_epiparameter_distribution(onset_to_death_ebola, by = 1)

}
