% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rolling_cfr.R
\name{rolling_cfr}
\alias{rolling_cfr}
\title{Rolling case fatality rate}
\usage{
rolling_cfr(
  df_in,
  correct_for_delays = TRUE,
  delay_pmf,
  poisson_threshold = 200
)
}
\arguments{
\item{df_in}{A data.frame containing the outbreak data. A daily time series
with dates or some other absolute indicator of time (e.g. epiday/epiweek) and
the numbers of new cases and new deaths at each time point}

\item{correct_for_delays}{A boolean flag indicating whether the user wishes
to correct for the delay between case detection and death. FALSE corresponds
to a naive CFR being calculated, TRUE corresponds to the user calculating a
corrected CFR}

\item{delay_pmf}{The delay distribution used, in the form of a probability
mass function parameterised by time. I.e. f(t) which gives the probability a
case has a known outcomes (i.e. death) at time t, parameterised with
disease-specific parameters before it is supplied here. A typical example
would be a symptom onset to death delay distribution}

\item{poisson_threshold}{The case count above which to use Poisson
approximation. Set to 200 by default.}
}
\value{
A data.frame containing the same case and death time series supplied
with the desired CFR (naive or corrected) appended at each time point
as extra columns
}
\description{
Calculates the CFR at each time point in the case and death time
series supplied, using an expanding window of time. The static CFR is
calculated for each time point, using the time series from the start to each
time point increasing the number of time points included by one each
iteration
}
\examples{
# Get onset to death distribution from epiparameter
onset_to_death_ebola <- epiparameter::epidist("ebola", "onset_to_death")
delay_pmf <- onset_to_death_ebola$pmf

# Calculate rolling naive CFR
df_ncfr <- rolling_cfr(ebola1976, correct_for_delays = FALSE)

# Calculate rolling corrected CFR
df_ccfr <- rolling_cfr(
  df_in = ebola1976,
  correct_for_delays = TRUE,
  delay_pmf
)
}
