% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/static_cfr.R
\name{static_cfr}
\alias{static_cfr}
\title{Static case fatality rate}
\usage{
static_cfr(df_in, correct_for_delays = TRUE, epi_dist, poisson_threshold = 200)
}
\arguments{
\item{df_in}{A data.frame containing the outbreak data. A daily time series
with dates or some other absolute indicator of time (e.g. epiday/epiweek) and
the numbers of new cases and new deaths at each time point}

\item{correct_for_delays}{A boolean flag indicating whether the user wishes
to correct for the delay between case detection and death. FALSE corresponds
to a naive CFR being calculated, TRUE corresponds to the user calculating a
corrected CFR}

\item{epi_dist}{The delay distribution used, in the form of an
\code{\link[epiparameter:epidist]{epiparameter::epidist()}} object. This is used to obtain a probability
mass function parameterised by time; i.e. \eqn{f(t)} which gives the
probability a case has a known outcomes (i.e. death) at time \eqn{t},
parameterised with disease-specific parameters before it is supplied here.
A typical example would be a symptom onset to death delay distribution.}

\item{poisson_threshold}{The case count above which to use Poisson
approximation. Set to 200 by default.}
}
\value{
A named vector with the MLE and 95\% confidence interval of the
CFR estimates, named "cfr_me", "cfr_low", and "cfr_high".
}
\description{
Calculates the CFR at a single point in time using in the case
and death time series supplied. Uses all of the time points supplied by the
user and returns an estimate for a single point in time.
}
\examples{
# Create an `epidist` for EVD onset to death distribution
# taken from parameters in 10.1016/S0140-6736(18)31387-4
onset_to_death_ebola <- epiparameter::epidist(
  disease = "Ebola virus disease",
  pathogen = "Ebolavirus",
  epi_dist = "onset_to_death",
  prob_distribution = "gamma",
  prob_distribution_params = c(
    shape = 2.4, scale = 3.333
  )
)

# load stored ebola case data from 1976 outbreak
data("ebola1976")

# Calculate static naive CFR using saved
static_cfr(df_in = ebola1976, correct_for_delays = FALSE)

# Calculate static corrected CFRs
static_cfr(
  df_in = ebola1976,
  correct_for_delays = TRUE,
  onset_to_death_ebola
)
}
